Consolidate libxml2 tests into main test file
Replace build constraints with runtime checks for Linux and libxml2
availability, removing the need for a separate test file
